LayoutTests/platform/efl/svg/dynamic-updates/SVGFESpotLightElement-svgdom-y-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spotLightElement.y.baseVal is 100
PASS spotLightElement.y.baseVal is 0
PASS successfullyParsed is true

TEST COMPLETE
